requirements.txt
openai
streamlit
Pillow
streamlit-lottie
mysql-connector-python
requests
toml